+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 
 ### 0.0.2
 - Changes names
+
+
+### 0.0.3
+- Allow additional query parameters in GET methods
+- Add static functions to generate select, filter, order, limit, skip parameters for GET methods
+- Cleanup